Extract hex_transaction module
